requirements.txt
# Requirements for Advanced Scalping Bot with Intelligent Features

# Core dependencies
binance-connector>=3.0.0
pandas>=1.5.0
numpy>=1.23.0
python-decouple>=3.8

# Intelligent Features
requests>=2.28.0  # For API calls (Fear & Greed Index, economic calendar)

# Web Server & API (for Railway deployment)
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
websockets>=12.0
pydantic>=2.5.0
python-multipart>=0.0.6
jinja2>=3.1.0
email-validator>=2.0.0

# Database (PostgreSQL on Railway)
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
alembic>=1.13.0

# Security & Auth
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==3.2.2
python-dotenv>=1.0.0
cryptography>=41.0.0

# Config Management
pyyaml>=6.0.1

# Testing (Development)
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
pytest-mock>=3.11.0
httpx>=0.24.0

# Optional (if you want to use ta-lib instead of manual calculations)
# TA-Lib>=0.4.0
